Quick-and-dirty Android app for "smart" watch
This is a no-name Chinese watch, whose app tested positive in VirusTotal
against several engines.

I was able to reverse-engineer two commands from the app to the watch,
one to "find" the watch, by making it make noise until the user presses
a button, and one to display a notification on the watch.

I took the Android sample Bluetooth Chat app as a base, added a
notification listener, and replaced the chat part with a "Find Watch"
button.